Title: 특정 연산을 이행하는 함수 2, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240319/특정 연산을 이행하는 함수 2/functions-that-perform-certain-operations-2.cpp
@@ -5,7 +5,7 @@ using namespace std;
 int main() {
     double a, b, c;
     cin >> a >> b >> c;
-    int tmp;
+    double tmp;
     if(a<c){
         tmp = a;
         a = c;
